refactor : useCheckedState 로 분리
